問題対処： マルチカーソル：選択状態での文字挿入が動作不良
git-svn-id: https://svn.osdn.net/svnroot/vivi-50-free/trunk@155 3a3fcbb5-9455-4931-b1bf-4cf1dfffbb8a

+++ vivi50/TextDocument.h
@@ -338,7 +338,7 @@ public:
 	void	do_erase(index_t, index_t, ushort=0);				//	undo/redo ‘Î‰ž”Å
 	void	do_replace(index_t, index_t, const QString &);		//	undo/redo ‘Î‰ž”Å
 
-	size_t	insertText(TextCursor&, const QString &, bool = false);
+	int		insertText(TextCursor&, const QString &, bool = false);
 	size_t	deleteChar(TextCursor&);
 	size_t	deletePreviousChar(TextCursor&);
