+++ vivi50/TextView.h
@@ -76,7 +76,7 @@ public:
 	void	deleteChar();
 	void	deletePreviousChar();
 	void	insertText(const QString &, bool = false);
-	size_t	insertText(ViewTextCursor &, const QString &);
+	int		insertText(ViewTextCursor &, const QString &);
 	void	deleteChar(ViewTextCursor &);
 	void	deletePreviousChar(ViewTextCursor &);
